database(schema): size is bigint now

+++ sql/mfs_schema.sql
@@ -2,7 +2,7 @@ CREATE TABLE mfs (
             PATH 	TEXT PRIMARY KEY UNIQUE,
             PARENT	TEXT NOT NULL,
             DIR		BOOLEAN NOT NULL,
-            SIZE 	INTEGER NOT NULL,
+            SIZE 	BIGINT NOT NULL,
             MODTIME TIME NOT NULL,
             KEY     TEXT,
             OWNER   TEXT
